feat(workspace-settings): add workspace settings
