+++ package.json
@@ -39,16 +39,16 @@
     "@fontsource-variable/plus-jakarta-sans": "^5.1.0",
     "@lingui/macro": "^4.13.0",
     "@lingui/react": "^4.13.0",
-    "@mui/icons-material": "^6.1.4",
-    "@mui/lab": "^6.0.0-beta.12",
-    "@mui/material": "^6.1.4",
-    "@mui/system": "^6.1.4",
+    "@mui/icons-material": "^6.1.5",
+    "@mui/lab": "^6.0.0-beta.13",
+    "@mui/material": "^6.1.5",
+    "@mui/system": "^6.1.5",
     "@mui/x-charts": "^7.21.0",
     "@mui/x-charts-pro": "^7.0.0-beta.5",
     "@mui/x-data-grid-premium": "^7.21.0",
     "@mui/x-date-pickers-pro": "^7.21.0",
     "@mui/x-tree-view": "^7.21.0",
-    "@tanstack/react-query": "^5.59.15",
+    "@tanstack/react-query": "^5.59.16",
     "@uidotdev/usehooks": "^2.4.1",
     "axios": "^1.7.7",
     "d3": "^7.9.0",
@@ -61,7 +61,7 @@
     "js-cookie": "^3.0.5",
     "lowlight": "^3.1.0",
     "mustache": "^4.2.0",
-    "posthog-js": "^1.174.2",
+    "posthog-js": "^1.175.0",
     "qrcode": "^1.5.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -95,7 +95,7 @@
     "@types/js-cookie": "^3.0.6",
     "@types/mustache": "^4.2.5",
     "@types/qrcode": "^1.5.5",
-    "@types/react": "^18.3.11",
+    "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/react-is": "^18.3.0",
     "@types/react-lazy-load-image-component": "^1.6.4",
@@ -122,15 +122,15 @@
     "tsc-files": "^1.1.4",
     "type-fest": "^4.26.1",
     "typescript": "^5.6.3",
-    "typescript-eslint": "^8.10.0",
+    "typescript-eslint": "^8.11.0",
     "unique-names-generator": "^4.7.1",
     "uuid": "^10.0.0",
-    "vite": "^5.4.9",
+    "vite": "^5.4.10",
     "vite-plugin-html": "^3.2.2",
     "vite-plugin-mock-dev-server": "^1.8.0",
     "vite-plugin-svgr": "^4.2.0",
     "vitest": "^2.1.3",
-    "web-vitals": "^4.2.3"
+    "web-vitals": "^4.2.4"
   },
   "browserslist": {
     "production": [
