+++ src/pages/panel/workspace-settings-connected-services/WorkspaceSettingsPagerdutyService.tsx
@@ -1,30 +1,29 @@
-import { Trans, t } from '@lingui/macro'
-import PowerIcon from '@mui/icons-material/Power'
-import SettingsIcon from '@mui/icons-material/Settings'
+import { t, Trans } from '@lingui/macro'
 import { LoadingButton } from '@mui/lab'
-import { Button, Stack, TextField, Typography, useTheme } from '@mui/material'
+import { alpha, Button, Stack, TextField, Typography, useTheme } from '@mui/material'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useEffect, useRef, useState } from 'react'
-import { OpsgenieLogo } from 'src/assets/icons'
+import { OpenInNewIcon, PagerdutyLogo, PowerIcon, SettingsIcon } from 'src/assets/icons'
 import { useUserProfile } from 'src/core/auth'
 import { Modal } from 'src/shared/modal'
+import { putWorkspaceNotificationAddMutation } from './putWorkspaceNotificationAdd.mutation'
+import { WorkspaceSettingsConnectedServiceItemContainer } from './WorkspaceSettingsConnectedServiceItemContainer'
 import { WorkspaceSettingsDisconnectServiceModal } from './WorkspaceSettingsDisconnectServiceModal'
 import { WorkspaceSettingsTestService } from './WorkspaceSettingsTestService'
-import { putWorkspaceNotificationAddMutation } from './putWorkspaceNotificationAdd.mutation'
 
-interface WorkspaceSettingsOpsgenieServiceProps {
+interface WorkspaceSettingsPagerdutyServiceProps {
   isConnected?: boolean
   defaultName?: string
   isLoading?: boolean
 }
 
-export const WorkspaceSettingsOpsgenieService = ({ isConnected, defaultName, isLoading }: WorkspaceSettingsOpsgenieServiceProps) => {
+export const WorkspaceSettingsPagerdutyService = ({ isConnected, defaultName, isLoading }: WorkspaceSettingsPagerdutyServiceProps) => {
   const theme = useTheme()
   const { selectedWorkspace, checkPermission } = useUserProfile()
   const hasPermission = checkPermission('updateSettings')
   const { mutate, isPending } = useMutation({ mutationFn: putWorkspaceNotificationAddMutation })
-  const [apiKey, setApiKey] = useState('')
-  const [name, setName] = useState(defaultName ?? 'Opsgenie Integration')
+  const [integrationKey, setIntegrationKey] = useState('')
+  const [name, setName] = useState(defaultName ?? 'Pagerduty Integration')
   useEffect(() => {
     if (defaultName) {
       setName(defaultName)
@@ -34,7 +33,7 @@ export const WorkspaceSettingsOpsgenieService = ({ isConnected, defaultName, isL
   const modalRef = useRef<(show?: boolean | undefined) => void>()
   const handleConnect = () => {
     mutate(
-      { workspaceId: selectedWorkspace?.id ?? '', api_key: apiKey, name, channel: 'opsgenie' },
+      { workspaceId: selectedWorkspace?.id ?? '', integration_key: integrationKey, name, channel: 'pagerduty' },
       {
         onSettled: () => {
           queryClient.invalidateQueries({
@@ -46,23 +45,14 @@ export const WorkspaceSettingsOpsgenieService = ({ isConnected, defaultName, isL
     )
   }
   return hasPermission || isConnected ? (
-    <Stack
-      direction={{ xs: 'column', sm: 'row' }}
-      alignItems={{ xs: 'stretch', sm: 'center' }}
-      justifyContent={{ xs: 'center', sm: 'start' }}
-      minHeight={40}
-      flexWrap="wrap"
-      gap={2}
-    >
-      <Stack width={150} justifyContent="center">
-        <OpsgenieLogo color={theme.palette.common.black} width={120} />
-      </Stack>
+    <WorkspaceSettingsConnectedServiceItemContainer icon={<PagerdutyLogo color={theme.palette.common.black} height={38} />}>
       {hasPermission ? (
         <>
-          {isConnected ? <WorkspaceSettingsDisconnectServiceModal isLoading={isLoading} channel="opsgenie" name="Opsgenie" /> : null}
+          {isConnected ? <WorkspaceSettingsDisconnectServiceModal isLoading={isLoading} channel="pagerduty" name="PagerDuty" /> : null}
           <LoadingButton
             loadingPosition={isLoading && !isConnected ? 'start' : undefined}
             startIcon={isConnected ? <SettingsIcon /> : <PowerIcon />}
+            color={isConnected ? 'primary' : 'success'}
             loading={isLoading}
             variant="contained"
             sx={{ flexShrink: 0 }}
@@ -70,37 +60,35 @@ export const WorkspaceSettingsOpsgenieService = ({ isConnected, defaultName, isL
           >
             {isConnected ? <Trans>Configure</Trans> : <Trans>Connect</Trans>}
           </LoadingButton>
-          {isConnected ? <WorkspaceSettingsTestService channel="opsgenie" isLoading={isLoading} /> : null}
+          {isConnected ? <WorkspaceSettingsTestService channel="pagerduty" isLoading={isLoading} /> : null}
           <Modal
+            slotProps={{ backdrop: { sx: { bgcolor: alpha('#000000', 0.6) } } }}
             onSubmit={handleConnect}
             openRef={modalRef}
-            title={<Trans>Connect Opsgenie</Trans>}
+            title={<Trans>Connect PagerDuty</Trans>}
             description={
               <Trans>
                 <Typography component="div">
-                  We utilize the OpsGenie API integration for notifications and alerts.
+                  We use the Events API V2 for notifications through PagerDuty.
                   <br />
-                  To set up this Integration, please follow these instructions:
+                  To create an integration, please follow these steps:
                   <ol>
-                    <li>
-                      <strong>Navigate to the Dashboard:</strong> Find and click on the “Teams” section from the sidebar once logged in.
-                    </li>
-                    <li>
-                      <strong>Select Your Team:</strong> Choose the team you want to integrate with. Within the team dashboard, go to the
-                      “Integrations” tab.
-                    </li>
-                    <li>
-                      <strong>Add Integration:</strong> Click on the “+ New Integration” button. Select “API” as the integration type, give
-                      it the name “Fix,” and press Continue.
-                    </li>
-                    <li>
-                      <strong>Turn Integration on:</strong> When the Integration has been created, it is disabled by default - click “Turn
-                      on Integration”.
-                    </li>
-                    <li>
-                      <strong>Copy API Key:</strong> Find and copy the API Key in the integration settings.
-                    </li>
+                    <li>Log in to PagerDuty with your account credentials.</li>
+                    <li>Navigate to the dashboard and select "Services".</li>
+                    <li>Choose the service you wish to integrate, go to the "Integrations" tab, and then click "Add Integration".</li>
+                    <li>For the integration type, select "PagerDuty Events API V2" and click "Add".</li>
+                    <li>Provide a name for the integration for easy reference and copy the "Integration Key".</li>
                   </ol>
+                  For more information please visit{' '}
+                  <Button
+                    size="small"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    href="https://support.pagerduty.com/docs/services-and-integrations#section-events-API-v2"
+                    endIcon={<OpenInNewIcon />}
+                  >
+                    PagerDuty Services and Integrations
+                  </Button>
                 </Typography>
               </Trans>
             }
@@ -113,18 +101,19 @@ export const WorkspaceSettingsOpsgenieService = ({ isConnected, defaultName, isL
                 )}
                 <LoadingButton
                   loadingPosition={isPending ? 'start' : undefined}
-                  startIcon={<PowerIcon />}
                   loading={isPending}
                   variant="contained"
                   type="submit"
-                  disabled={!name || !apiKey}
+                  disabled={!name || !integrationKey}
+                  startIcon={isConnected ? <SettingsIcon /> : <PowerIcon />}
+                  color={isConnected ? 'primary' : 'success'}
                 >
-                  <Trans>Connect</Trans>
+                  {isConnected ? <Trans>Update</Trans> : <Trans>Connect</Trans>}
                 </LoadingButton>
               </>
             }
           >
-            <Stack spacing={2}>
+            <Stack spacing={2} my={2}>
               <TextField
                 required
                 name="name"
@@ -138,20 +127,19 @@ export const WorkspaceSettingsOpsgenieService = ({ isConnected, defaultName, isL
               />
               <TextField
                 required
-                name="api_key"
-                autoComplete="api_key"
-                label={t`API Key`}
+                name="integration_key"
+                autoComplete="text"
+                label={t`Integration Key`}
                 variant="outlined"
                 fullWidth
                 type="text"
-                value={apiKey}
-                onChange={(e) => setApiKey(e.target.value ?? '')}
-                autoFocus
+                value={integrationKey}
+                onChange={(e) => setIntegrationKey(e.target.value ?? '')}
               />
             </Stack>
           </Modal>
         </>
       ) : null}
-    </Stack>
+    </WorkspaceSettingsConnectedServiceItemContainer>
   ) : null
 }
