+++ src/pages/panel/workspace-settings-connected-services/WorkspaceSettingsSlackService.tsx
@@ -1,49 +1,39 @@
 import { Trans } from '@lingui/macro'
-import PowerIcon from '@mui/icons-material/Power'
 import { LoadingButton } from '@mui/lab'
-import { Stack, useTheme } from '@mui/material'
-import { DiscordWithTextLogo } from 'src/assets/icons'
+import { useTheme } from '@mui/material'
+import { PowerIcon, SlackWithTextLogo } from 'src/assets/icons'
 import { useUserProfile } from 'src/core/auth'
 import { endPoints } from 'src/shared/constants'
+import { WorkspaceSettingsConnectedServiceItemContainer } from './WorkspaceSettingsConnectedServiceItemContainer'
 import { WorkspaceSettingsDisconnectServiceModal } from './WorkspaceSettingsDisconnectServiceModal'
 import { WorkspaceSettingsTestService } from './WorkspaceSettingsTestService'
 
-interface WorkspaceSettingsDiscordServiceProps {
+interface WorkspaceSettingsSlackServiceProps {
   isConnected?: boolean
   isLoading?: boolean
 }
 
-export const WorkspaceSettingsDiscordService = ({ isConnected, isLoading }: WorkspaceSettingsDiscordServiceProps) => {
+export const WorkspaceSettingsSlackService = ({ isConnected, isLoading }: WorkspaceSettingsSlackServiceProps) => {
   const { selectedWorkspace, checkPermission } = useUserProfile()
   const hasPermission = checkPermission('updateSettings')
 
   const theme = useTheme()
   return hasPermission || isConnected ? (
-    <Stack
-      direction={{ xs: 'column', sm: 'row' }}
-      alignItems={{ xs: 'stretch', sm: 'center' }}
-      justifyContent={{ xs: 'center', sm: 'start' }}
-      minHeight={40}
-      flexWrap="wrap"
-      gap={2}
-    >
-      <Stack width={150} justifyContent="center">
-        <DiscordWithTextLogo color={theme.palette.common.black} width={120} />
-      </Stack>
+    <WorkspaceSettingsConnectedServiceItemContainer icon={<SlackWithTextLogo color={theme.palette.common.black} height={38} />}>
       {hasPermission ? (
         isConnected ? (
           <>
-            <WorkspaceSettingsDisconnectServiceModal channel="discord" isLoading={isLoading} name="Discord" />
-            <WorkspaceSettingsTestService channel="discord" isLoading={isLoading} />
+            <WorkspaceSettingsDisconnectServiceModal isLoading={isLoading} channel="slack" name="Slack" />
+            <WorkspaceSettingsTestService channel="slack" isLoading={isLoading} />
           </>
         ) : (
           <LoadingButton
             loadingPosition={isLoading ? 'start' : undefined}
             startIcon={<PowerIcon />}
-            href={endPoints.workspaces.workspace(selectedWorkspace?.id ?? '').notification.add('discord')}
+            href={endPoints.workspaces.workspace(selectedWorkspace?.id ?? '').notification.add('slack')}
             loading={isLoading}
             variant="contained"
-            color="primary"
+            color="success"
             sx={{ flexShrink: 0 }}
             disabled={!selectedWorkspace?.id}
           >
@@ -51,6 +41,6 @@ export const WorkspaceSettingsDiscordService = ({ isConnected, isLoading }: Work
           </LoadingButton>
         )
       ) : null}
-    </Stack>
+    </WorkspaceSettingsConnectedServiceItemContainer>
   ) : null
 }
