--- src/pages/panel/workspace-settings/workspace-alerting-settings/index.ts
@@ -1,1 +1,0 @@
-export { WorkspaceAlertingSettings } from './WorkspaceAlertingSettings'
