--- src/pages/panel/workspace-settings/workspace-settings-services/index.ts
@@ -1,1 +1,0 @@
-export { WorkspaceSettingsConnectedServices } from './WorkspaceSettingsConnectedServices'
