+++ src/shared/severity/index.ts
@@ -1,0 +1,2 @@
+export { SeverityDropDown } from './SeverityDropDown'
+export { SeverityItem } from './SeverityItem'
